+++ src/stack.c
@@ -2,7 +2,7 @@
 
 #include "calc.h"
 // положить число в стек
-void push(create_stack *stack, int value) {
+void push(create_stack *stack, double value) {
   if (stack->size >= STACK_MAX_SIZE) {
     printf("STACK_OVERFLOW");
     exit(STACK_OVERFLOW);
@@ -15,7 +15,7 @@ void push(create_stack *stack, int value) {
 }
 
 // взять число из стека
-int pop(create_stack *stack) {
+double pop(create_stack *stack) {
   if (stack->size == 0) {
     printf("STACK_UNDERFLOW pop");
     exit(STACK_UNDERFLOW);
@@ -25,7 +25,7 @@ int pop(create_stack *stack) {
 }
 
 // посмотреть последнее число из стека
-int peek(const create_stack *stack) {
+double peek(const create_stack *stack) {
   if (stack->size == 0) {
     printf("STACK_UNDERFLOW peek");
     exit(STACK_UNDERFLOW);
